+++ R/CFAxisScalar.R
@@ -13,18 +13,22 @@ CFAxisScalar <- R6::R6Class("CFAxisScalar",
   inherit = CFAxis,
   private = list(
     dimvalues_short = function() {
-      paste0("[", self$value, "]")
+      v <- if (inherits(self$value, "CFTime")) as_timestamp(self$value)
+           else self$value
+      paste0("[", v, "]")
     }
   ),
   public = list(
-    #' @field value The value of the axis.
+    #' @field value The value of the axis. This could be a composite value, such
+    #' as a `CFTime` instance.
     value = NULL,
 
     #' @description Create an instance of this class.
     #' @param grp The group that contains the netCDF variable.
     #' @param nc_var The netCDF variable that describes this instance.
     #' @param orientation The orientation of this axis, or "" if not known.
-    #' @param value The value of this axis.
+    #' @param value The value of this axis, possibly a compound type like
+    #' `CFTime`.
     initialize = function(grp, nc_var, orientation, value) {
       dim <- NCDimension$new(-1L, nc_var$name, 1L, FALSE)
       super$initialize(grp, nc_var, dim, orientation)
@@ -45,12 +49,19 @@ CFAxisScalar <- R6::R6Class("CFAxisScalar",
 
       cat("Axis     :", self$orientation, "\n")
 
-      units <- self$attribute("units")
-      if (!nzchar(units)) units <- ""
-      cat("Value    : ", self$value, " ", units, "\n", sep = "")
-      if (inherits(self$bounds, "CFBounds"))
-        self$bounds$print()
-      else cat("Bounds   : (not set)\n")
+      if (inherits(self$value, "CFTime")) {
+        cat("Value    :", as_timestamp(self$value), "\n")
+        bnds <- self$value$get_bounds("timestamp")
+        if (is.null(bnds)) cat("Bounds   : (not set)\n")
+        else cat("Bounds   : ", bnds[1L, 1L], ", ", bnds[2L, 1L], "\n", sep = "")
+      } else {
+        units <- self$attribute("units")
+        if (!nzchar(units)) units <- ""
+        cat("Value    : ", self$value, " ", units, "\n", sep = "")
+        if (inherits(self$bounds, "CFBounds"))
+          self$bounds$print()
+        else cat("Bounds   : (not set)\n")
+      }
 
       self$print_attributes()
     },
@@ -92,7 +103,8 @@ CFAxisScalar <- R6::R6Class("CFAxisScalar",
     #' @field dimnames (read-only) The coordinate of the axis.
     dimnames = function(value) {
       if (missing(value))
-        self$value
+        if (inherits(self$value, "CFTime")) as_timestamp(self$value)
+        else self$value
     }
   )
 )